issue: 3414980 Move applications code in separate file
