+++ src/core/sock/sock-app.h
@@ -1,0 +1,126 @@
+/*
+ * Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ *
+ * This software is available to you under a choice of one of two
+ * licenses.  You may choose to be licensed under the terms of the GNU
+ * General Public License (GPL) Version 2, available from the file
+ * COPYING in the main directory of this source tree, or the
+ * BSD license below:
+ *
+ *     Redistribution and use in source and binary forms, with or
+ *     without modification, are permitted provided that the following
+ *     conditions are met:
+ *
+ *      - Redistributions of source code must retain the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer.
+ *
+ *      - Redistributions in binary form must reproduce the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer in the documentation and/or other materials
+ *        provided with the distribution.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#ifndef _SOCK_APP_H_
+#define _SOCK_APP_H_
+
+#include <core/util/sys_vars.h>
+#include <core/util/utils.h>
+#include <utils/lock_wrapper.h>
+
+#include <vlogger/vlogger.h>
+#include <unordered_map>
+#include <mutex>
+#include <set>
+
+#if defined(DEFINED_NGINX)
+typedef std::unordered_map<uint16_t, bool> map_udp_bounded_port_t;
+
+extern map_udp_bounded_port_t g_map_udp_bounded_port;
+#endif
+
+#if defined(DEFINED_NGINX) || defined(DEFINED_ENVOY)
+struct app_conf {
+    app_type_t type;
+    lock_spin_recursive m_lock;
+    int workers_num;
+    int workers_pow2;
+    int src_port_stride;
+    bool add_second_4t_rule;
+    /* Map listen fd with thread(process) */
+    std::unordered_map<int, pid_t> map_listen_fd;
+    /* Associate thread(process) with unique identifier */
+    std::unordered_map<pid_t, int> map_thread_id;
+    /* Store duplicated fdwith related original fd */
+    std::unordered_map<int, int> map_dup_fd;
+    /* Collection of unused unique identifiers limited by workers_num */
+    std::set<int> unused_worker_id;
+    void *context;
+
+    app_conf()
+    {
+        type = APP_NONE;
+        m_lock = lock_spin_recursive("app_conf");
+        workers_num = 0;
+        workers_pow2 = 0;
+        src_port_stride = 2;
+        add_second_4t_rule = false;
+        map_listen_fd.clear();
+        map_thread_id.clear();
+        map_dup_fd.clear();
+        unused_worker_id.clear();
+        context = NULL;
+
+        setup();
+    }
+
+    ~app_conf() {}
+
+    void setup()
+    {
+        type = safe_mce_sys().app.type;
+        workers_num = safe_mce_sys().app.workers_num;
+        src_port_stride = safe_mce_sys().app.src_port_stride;
+
+        // Round up to a power of 2 value. Assume the number doesn't exceed 32bit.
+        workers_pow2 = workers_num - 1;
+        workers_pow2 |= workers_pow2 >> 1;
+        workers_pow2 |= workers_pow2 >> 2;
+        workers_pow2 |= workers_pow2 >> 4;
+        workers_pow2 |= workers_pow2 >> 8;
+        workers_pow2 |= workers_pow2 >> 16;
+        workers_pow2++;
+    }
+
+    inline int get_worker_id()
+    {
+        std::lock_guard<decltype(this->m_lock)> lock(m_lock);
+        auto itr = map_thread_id.find(gettid());
+        if (itr != map_thread_id.end()) {
+            return itr->second;
+        }
+        return -1;
+    }
+
+#if defined(DEFINED_NGINX)
+    int proc_nginx(void);
+#endif /* DEFINED_NGINX */
+
+#if defined(DEFINED_ENVOY)
+    int proc_envoy(int op, int fd);
+#endif /* DEFINED_ENVOY */
+};
+
+extern struct app_conf *g_p_app;
+#endif
+
+#endif /* _SOCK_APP_H_ */
